# light culling completed for point lights