--- a/Engine/Game/Assets/Shaders/TileLightCulling.comp
+++ b/Engine/Game/Assets/Shaders/TileLightCulling.comp
@@ -20,6 +20,10 @@ layout(location = 0)uniform uint maxPointLights;
 //layout(location = 1)uniform uvec2 numPixels;
 shared uint pointsInTile;
 
+uniform layout(binding = 1, r32f) readonly image2D depthImage;
+//layout (binding = 0) uniform sampler2D gDepth;
+shared uint minZ;
+shared uint maxZ;
 
 float GetLinearZ(float inputDepth)
 {
@@ -46,9 +50,16 @@ void main()
 
 	if(gl_LocalInvocationIndex == 0)
 	{
+		minZ = 0xFFFFFFFF;
+        maxZ = 0;
 		pointsInTile = 0;
 	}
 	barrier();
+
+	uint pixelDepth = floatBitsToUint(imageLoad(depthImage, ivec2(gl_GlobalInvocationID.xy)).r);
+	//uint pixelDepth = floatBitsToUint(texture2D(gDepth, ivec2(vec2(gl_GlobalInvocationID.xy) / vec2(numPixels))).r);
+	atomicMin(minZ, pixelDepth);
+	atomicMax(maxZ, pixelDepth);
 	
 	vec3 farPlanePoints[4];
 	farPlanePoints[0] = GetViewPos(1.0, vec2(float(gl_WorkGroupID.x)/float(gl_NumWorkGroups.x),   float(gl_WorkGroupID.y)/float(gl_NumWorkGroups.y)));
@@ -61,7 +72,11 @@ void main()
 	{
 		tileFrustumPlanes[i] = GetFrustumPlane(farPlanePoints[i], farPlanePoints[(i+1)&3]);
 	}
-	
+
+	barrier();
+	float viewMinZ = GetLinearZ(uintBitsToFloat(minZ));
+	float viewMaxZ = GetLinearZ(uintBitsToFloat(maxZ));
+
 	for(uint i=gl_LocalInvocationIndex;i<numPLights; i+=threadCount)
 	{
 		PointLight light = pLights[i];
@@ -69,7 +84,9 @@ void main()
 		if(dot(viewPos, tileFrustumPlanes[0].xyz) < light.pos.w &&
 			dot(viewPos, tileFrustumPlanes[1].xyz) < light.pos.w &&
 			dot(viewPos, tileFrustumPlanes[2].xyz) < light.pos.w &&
-			dot(viewPos, tileFrustumPlanes[3].xyz) < light.pos.w)
+			dot(viewPos, tileFrustumPlanes[3].xyz) < light.pos.w &&
+			(viewMinZ - viewPos.z) > (-light.pos.w) &&
+            (viewMaxZ - viewPos.z) < light.pos.w)
 		{
 			uint index = atomicAdd(pointsInTile, 1);
 			if(index < maxPointLights)