--- a/Engine/Source/ModuleDebugDraw.cpp
+++ b/Engine/Source/ModuleDebugDraw.cpp
@@ -635,7 +635,7 @@ update_status  ModuleDebugDraw::Update(float dt)
         {
             float4x4 viewproj = camera->GetProjectionMatrix() * camera->GetViewMatrix();
             
-            EngineApp->GetOpenGL()->BindGFramebuffer();
+            EngineApp->GetOpenGL()->BindSceneFramebuffer();
             Draw(viewproj, EngineApp->GetWindow()->GetGameWindowsSize().x, EngineApp->GetWindow()->GetGameWindowsSize().y);
             EngineApp->GetOpenGL()->UnbindFramebuffer();
         }