… Engine/Source/ModuleResource.cpp