finished the tile light culling adding the near and far plane on the tiles frustums

results: (always depending where we are and how many lights are on the zone inside the gameplay camera)
the scene right now has 169 point lights and a 2k screen
without tile light culling: 
9-10 ms (aprox 100fps)  lighting pass costs 6,6ms per frame
with tile light culling:
5 ms (200 fps) lighting pass costs 1,94ms per frame, adding a compute shader cost of 0,06ms
Notes:
We are two times faster:
havent't observed a lot of change with the near and far plane constructing the tiles frustum using the depth buffer but i supose that it depends of what is inside the camera, our camera is facing the floor so we don't have a lot of underground lights to cull
TODO:
Now the most costly is the generate shadow maps 3,5 ms and then the lighting pass with 1,94ms
Still missing the tilelights culling